Support an alternative, shell friendly, way to define ES settings
Shells allow a specific character class for environment variables:
`[a-zA-Z_]+[a-zA-Z0-9_]*`.  While Docker and POSIX[1] can deal with
dots in environment names, Kubernetes is currently also affected[2] by
this and as Elasticsearch settings contain dots, this creates
frustration for users who prefer customizing Elasticsearch using
environment vars.

This commit offers the possibility for users to define settings using
a schema prefixed with `ess__` and `__` instead of dots, like
`ess__cluster__name=my-es-on-docker`.

As Elasticsearch also uses _ in some env vars (like
`bootstrap.memory_lock`) and other env vars like `ES_JAVA_OPTS` are
already prefixed with `ES_` we have to resort to the prefix `ess__`
and using `__` as a delimiter.

[1]
http://pubs.opengroup.org/onlinepubs/009695399/basedefs/xbd_chap08.html
[2]
https://github.com/kubernetes/kubernetes/issues/2707#issuecomment-285309156